fix(fullscreen): fix after review add plugin/selector-bem-pattern for stylelint
