+++ packages/client/src/components/game/game.pcss
@@ -13,7 +13,7 @@
     flex: 0 0 60px;
   }
 
-  &__info-fullscreen {
+  &__info_fullscreen {
     width: 60vw;
     flex: 0 0 60px;
   }
@@ -22,12 +22,12 @@
     width: 60%;
   }
 
-  &__board-fullscreen {
+  &__board_fullscreen {
     width: 45vw;
   }
 }
 
-.game-fullscreen {
+.game_fullscreen {
   display: flex;
   flex-direction: column;
   align-items: center;
