+++ packages/server/package.json
@@ -31,6 +31,7 @@
     "jest": "^28",
     "nodemon": "^2.0.19",
     "prettier": "^2.7.1",
+    "stylelint-selector-bem-pattern": "^3.0.0",
     "ts-jest": "^28.0.8",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.2"
